+++ cogs/faceshifter/__init__.py
@@ -1,6 +1,0 @@
-from .faceshifter import Faceshifter
-
-
-def setup(bot):
-    """Load cog"""
-    bot.add_cog(Faceshifter(bot))